requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
httpx==0.25.2
weasyprint==60.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
websockets==11.0.3
pdfplumber==0.11.4
